+++ src/validators/string/string.validator.js
@@ -1,14 +1,17 @@
 const _ = require('lodash')
+const { getCurrentTranslation } = require('../../translation')
 
 const stringValidator = {
   $type: 'string',
   validate: ctx => {
     if (ctx.handlePreValidation().handled) return
 
+    const translation = getCurrentTranslation()
+
     if (!_.isString(ctx.value)) {
       ctx.registerErrors({
         type: {
-          message: 'The value should be a string',
+          message: translation.validators.string.errors.type,
           value: ctx.value
         }
       })
@@ -20,7 +23,7 @@ const stringValidator = {
     if (_.isNumber(ctx.schema.minLength)) {
       if (ctx.value.length < ctx.schema.minLength) {
         errors.minLength = {
-          message: `The min length for this field is ${ctx.schema.minLength}`,
+          message: translation.validators.string.errors.minLength(ctx),
           minLength: ctx.schema.minLength,
           currentLength: ctx.value.length,
           value: ctx.value
@@ -31,7 +34,7 @@ const stringValidator = {
     if (_.isNumber(ctx.schema.maxLength)) {
       if (ctx.value.length > ctx.schema.maxLength) {
         errors.maxLength = {
-          message: `The max length for this field is ${ctx.schema.maxLength}`,
+          message: translation.validators.string.errors.maxLength(ctx),
           maxLength: ctx.schema.maxLength,
           currentLength: ctx.value.length,
           value: ctx.value
@@ -42,7 +45,7 @@ const stringValidator = {
     if (_.isRegExp(ctx.schema.regex)) {
       if (!ctx.schema.regex.test(ctx.value)) {
         errors.regex = {
-          message: 'The value do not match the rules.',
+          message: translation.validators.string.errors.regex,
           regex: ctx.schema.regex,
           value: ctx.value
         }