+++ CHANGELOG.md
@@ -6,6 +6,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.2.0] - 2019-01-25
+
+### Added
+
+- Adding `email` validator.
+- Adding translations
+
+### Fixed
+
+- Fixing `boolean` validator link.
+
+
 ## [1.1.0] - 2019-01-25
 
 - First release